+++ client/app/components/NavigationBar.jsx
@@ -2,25 +2,8 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import DropdownMenu from './DropdownMenu';
 import Link from './Link';
+import Breadcrumbs from './Breadcrumbs';
 import PerformanceDegradationBanner from './PerformanceDegradationBanner';
-import { Route } from 'react-router-dom';
-
-const getElementsWithBreadcrumbs = (element) => {
-  if (!element.props.children) {
-    return [];
-  }
-
-  return React.Children.toArray(element.props.children).reduce((acc, child) => {
-    if (child.props.breadcrumb) {
-      return [...acc, {
-        path: child.props.path,
-        breadcrumb: child.props.breadcrumb
-      }];
-    }
-
-    return [...acc, ...getElementsWithBreadcrumbs(child)];
-  }, []);
-};
 
 export default class NavigationBar extends React.Component {
   render() {
@@ -29,17 +12,6 @@ export default class NavigationBar extends React.Component {
       dropdownUrls,
       userDisplayName
     } = this.props;
-
-    const breadcrumbComponents = getElementsWithBreadcrumbs(this).map(
-      (route) => <Route path={route.path} render={
-        (props) => <span>
-            <h2 id="page-title" className="cf-application-title">&nbsp; > &nbsp;</h2>
-            <Link id="cf-logo-link" to={props.match.url}>
-              <h2 id="page-title" className="cf-application-title">{route.breadcrumb}</h2>
-            </Link>
-          </span>
-        } />
-      );
 
     return <div><header className="cf-app-header">
         <div>
@@ -51,7 +23,9 @@ export default class NavigationBar extends React.Component {
                   <h2 id="page-title" className="cf-application-title">&nbsp; {appName}</h2>
                 </Link>
               </h1>
-              {breadcrumbComponents}
+              <Breadcrumbs>
+                {this.props.children}
+              </Breadcrumbs>
             </span>
             <span className="cf-dropdown cf-push-right">
               <DropdownMenu
